+++ CHANGELOG.md
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Fixed
+
+- Nested conditional directives no longer take exponential time to parse.
+
 ## [0.5.0] - 2025-03-31
 
 ### Changed